Cleanup, refactor
