+++ apps/dapp/src/components/Pages/Ascend/components/AuctionChart.tsx
