--- apps/dapp/src/components/Pages/Auction/use-balancer-vault.ts
@@ -1,1 +1,0 @@
-import { Contract } from 'ethers';
